Is now $(pkglibdir)/include. `gnomesupport.h' is a generated file
* Makefile.am (supportincludedir): Is now $(pkglibdir)/include.
`gnomesupport.h' is a generated file dependent on the configure
process.  So, this seems a good place.  The rest of the headers
are also here just because I'm too lazy to maintain two separate
include directories to search in for `libgnomesupport' stuff ;-)
(supportinclude_DATA): New var.  Has `gnomesupport.h'.
(BUILT_SOURCES): Generate `gnomesupport.h'.

* gnomesupport-h.c: New file.  Used to generate `gnomesupport.h',
which contains prototypes for many of the functions in
`libgnomesupport'.  This method is used to ensure that prototypes
appear only for those functions that are provided by
`libgnomesupport'.

* scandir.c (alphasort): New utility function.

svn path=/trunk/; revision=190