Add lots of warnings to GNOME_COMPILE_WARNINGS
Following Colin Walters' thread on desktop-devel-list:

https://mail.gnome.org/archives/desktop-devel-list/2012-July/msg00100.html

Add a selection of more strict warnings to the default level of
GNOME_COMPILE_WARNINGS. Check all the warnings by passing them to GCC,
and ignore those which fail. Set the severity of some of the warnings to
errors, so that particularly bad code is rejected (set the warning level
to ‘minimum’ is you wish to avoid this behaviour). Fixes bug 568546 and
bug 608953.