macros2: Deprecate GNOME_COMPILE_WARNINGS
Use AX_COMPILER_FLAGS instead, which comes from the autoconf-archive. It
supports more compilers than just GCC, has a wider selection of flags,
supports LDFLAGS, and is maintained for more projects than just GNOME.

GNOME_COMPILE_WARNINGS has not been ported to AX_COMPILER_FLAGS as:
 • the default argument is a little difficult to map to
   AX_COMPILER_FLAGS’ IS-RELEASE argument; and
 • AX_COMPILER_FLAGS depends on some other macros in autoconf-archive
   licenced as GPLv3, which would potentially be tricky to copy into
   gnome-common.

The --enable-iso-c argument is deprecated as well. Projects should
choose which C standard they want to use, and hard-code that in their
compiler flags. It is not something which should be set by the
developer.

GNOME_CXX_WARNINGS is not deprecated, but may be in future.

No further changes should be made to GNOME_COMPILE_WARNINGS: please make
them to AX_COMPILER_FLAGS instead.

https://bugzilla.gnome.org/show_bug.cgi?id=729407