build: Make autoconf-archive support optional
As autoconf-archive installs the same ax_*.m4 macros that gnome-common
installs, which are copied from autoconf-archive into gnome-common, add
a configure argument to toggle autoconf-archive support.

Passing --without-autoconf-archive to configure (or not providing an
argument) disables compatibility (parallel-installability) with
autoconf-archive, and installs the ax_*.m4 macros. If
--with-autoconf-archive is passed to configure, the macros are not
installed, and gnome-common will not install any M4 macros that conflict
with autoconf-archive.

https://bugzilla.gnome.org/show_bug.cgi?id=747920